Add method GetJsonPaht
