+++ Lesson1Test2/Startup.cs
@@ -19,7 +19,6 @@ namespace Lesson1Test2
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddControllers();
-            //services.AddSingleton<FirstControler>();
             services.AddSingleton<ValuesHolder>();
         }
